+++ targets/frontend/.env
@@ -6,7 +6,7 @@
 ACCOUNT_MAIL_SENDER=contact@fabrique.social.gouv.fr
 
 # JWT lifetime (15min)
-JWT_TOKEN_EXPIRES=1
+JWT_TOKEN_EXPIRES=15
 
 # Refresh token lifetime (30 days in minutes)
 REFRESH_TOKEN_EXPIRES=43200
